Empirical-Risk-Minimization-
Empirical Risk Minimization" is a concept in the field of machine learning that focuses on developing automated models capable of performing well on the data they were trained on. This is achieved by minimizing the loss or gap between the predictions made by the models and the actual values in the training dataset. By doing so, the models aim to generalize well to new, unseen data, thus improving their overall performance and reliability.